+++ src/domain/bestand/ArtikelService.java
@@ -2,11 +2,11 @@ package domain.bestand;
 
 public class ArtikelService {
     //anlegen neuer Artikel
-    public Artikel artikelAnlegen(String kategorie_name, int id, String name, int bestand, int preis) {
+    public Artikel artikelAnlegen(String kategorie_name, int id, String name, int bestand, float preis) {
         return new Artikel(kategorie_name, id, name, bestand, preis);
     }
     //bearbeiten bereits existierender Artikel
-    public void artikelBearbeiten(Artikel artikel, String name, int bestand, int preis) {
+    public void artikelBearbeiten(Artikel artikel, String name, int bestand, float preis) {
         artikel.setArtikel_name(name);
         artikel.setBestand(bestand);
         artikel.setPreis(preis);
